Fix enyo.constructorForKind for enyo.getPath changes
enyo.getPath now treats "enyo" as the same as the global scope for lookups.  This caused a kind
lookup of "canvas.Control" to fail, since it was checking window.canvas.Control, not enyo.canvas.Control.
To fix, we change the logic so it does a string prepend of "enyo." to the kind string before calling
enyo.getPath instead of trying to scope to enyo object.

Enyo-DCO-1.1-Signed-Off-By: Ben Combee (ben.combee@lge.com)