Make enyo.Scroller provide bounds in events
While developing the new DataList, we saw a problem that the call get getScrollBounds
would cause animation to stop with the touch scroller.  Rather than change the behavior,
we'll instead use an internal method of the scroll strategies to get current bounds
and attach that structure to the onScroll, onScrollStart, and onScrollEnd events so users
of those events don't have to make calls back to the scroller.

Updated VerticalDelegate to use that info.  This is the only code where I directly saw
this problem in our current set of active libraries.

Enyo-DCO-1.1-Signed-Off-By: Ben Combee (ben.combee@lge.com)