Add check/box to tests title to indicate state.
This modified the tests.js test code file to add a checkmark
or box Unicode character to the start of the documen title when
we know if a test run has succeeded or failed.  This matches
when QUnit does, and only runs when we're not using QUinit as
our test runner.

Enyo-DCO-1.1-Signed-Off-By: Ben Combee (ben.combee@lge.com)