Fix this.inherited argument handling
Due to a rewrite of enyo.mixin, it lost the ability to combine
two plain arrays or an array-like object and an array.  Instead,
we'll inline simple code to do that in enyo.kind.inherited to
restore old functionality.

Also, add a test to validate this to KindTest.js in core tests.

Enyo-DCO-1.1-Signed-Off-By: Ben Combee (ben.combee@lge.com)