GF-40378: don't defer enyo.Signals due to early use
Code in cordova.js binds DOM event handlers to enyo.Signals.send
during Enyo load time. This binding would work, except that the
listeners variable modified by the call is in the protectedStatics
section, so it isn't created on the kind until it's undeferred.

This kind is small, so we'll just go ahead and mark it noDefer
instead.

Enyo-DCO-1.1-Signed-Off-By: Ben Combee (ben.combee@lge.com)