Replace replace() calls with enyo.trim()
The new observer and computed code uses strings of space-delimited
names instead of arrays for speed and memory savings, but the code
requires frequent operations to strip off spaces that may have
accumulated at the start and end.  ES5 provides a trim() method for
this, and tests using jsperf show it's twice the speed of the
equivalent replace() call in V8 and JSCore.

This also fixes latent bugs, as many of the replace() calls were
missing the global operator to strip off spaces both at start and
end.

Enyo-DCO-1.1-Signed-Off-By: Ben Combee (ben.combee@lge.com)