ENYO-3538: fix setObject handling of unchanged props
The enyo.Model.setObject code didn't correctly detect when a model
changed if the last property set from the new object was the same
as the original value, because a flag was getting overwritten for
each property change. This modifies this setObject code to notice
any change, instead of just the last one. A test is added to
core tests that fails on original code and passes now.

Enyo-DCO-1.1-Signed-Off-By: Ben Combee (ben.combee@lge.com)