ENYO-3534: fix touch events in IE10
MSPointerEvents events on IE10 on WinPhone8 have the buttons field set to 0.
Before, we hard coded the "which" value of the Enyo event to 1 and it worked
on the devices, but a PR to address ENYO-2993 made in August 2013 fixed desktop
issues but broke touch handling.  This was all masked when we added proper
pointer event support for IE11, but it left IE10 in the broken state.

Enyo-DCO-1.1-Signed-Off-By: Ben Combee (ben.combee@lge.com)