Update cached position for setScrollTop/Left when hidden
enyo.Scroller caches the scroll position when the scroller is
hidden or re-rendered.  Currently, API-level changes to the
scroll position made when the scroller is hidden are lost.
This change updates enyo.Scroller.setScrollTop/Left to update
the cache if it exists.

Enyo-DCO-1.1-Signed-Off-By: Ben Combee (ben.combee@lge.com)