ENYO-3852: fix loader handling of //-prefixed URLs
Moonstone and spotlight aren't loading for jsFiddle because its
loading their scripts via URLs that start with '//' instead of
'http://' in an attempt to future-proof for HTTPS or HTTP
deployment. Our loader mangled these in determining paths,
so we'll work around that be detecting this before parsing and
adding the protocol ourselves based on the current document.

Enyo-DCO-1.1-Signed-Off-By: Ben Combee (ben.combee@lge.com)