ENYO-3901: fix placeholder stretching and display
Placeholders for images are specified by an inline SVG
imgae. However, the height/width isn't specified by the SVG,
and the placeholder test only specified one of those, resulting
in a stretched appearance.

Also, some browsers had issue with the inline SVG XML
in the data URL, so replace it with the same code, but encoded as
base64.

Enyo-DCO-1.1-Signed-Off-By: Ben Combee (ben.combee@lge.com)