originally from 6eb48ddf154530c6cabb5a2cd73abc6cfabc96cc
Update the mouse coords in events when the body has been scaled. (Netcast 4)
Adds a new set of properties for calculating the clientX and clientY coordinates that are included on enyo events, which accurately reflect the position of the mouse on the screen.
The scale factors for X and Y are updated on load and on window resize.
Enyo-DCO-1.1-Signed-off-by: Blake Stephens <blake.stephens@lge.com>

Conflicts:
	source/dom/Control.js