originally from a9df0e234a652788f8c52ce85bb3c64c8dd02c3e
BHV-6058: Use absolute positioning to implement .enyo-scrollee-fit when scroller is used as a child of flexbox, since percentage-based sizing is not supported by flexbox.  According to notes in TouchScrollStrategy, the percentage sizing was only needed for IE to avoid a bug with non-zero scrollTop and abs pos siblings; since older versions of IE won't opt-in to the flexbox implementation, we can get away with reverting back to an enyo-fit implementation when using flexbox.
DCO-1.1-Signed-Off-By: Kevin Schaaf <kevin.schaaf@lge.com>