originally from 13dc2b4a5f6d563765071d0e2ccd34ca6c606d54
Support sync loading of binary files via enyo.Ajax

If you attempted to load binary files with enyo.Ajax synchronously,
the code kept trying to access xhr.responseText which is not a valid
thing to do because the response comes in xhr.response instead. This
would throw an error and make it impossible to load any binary file
synchronously.