Call start only when scroll position is actually changing.
There is unnecessary onScrollStart call in DataGridList render time.
DataGridList is reset itself when it is rendered and call scrollTo(0,0).
The scrollTo is calling doScrollStart without checking the position changes.
So, it is fires onScrollStart but there is no onScrollStop because it is not actually scrolling.

DCO-1.1-Signed-Off-By: Kunmyon Choi kunmyon.choi@lge.com