[Hotfix] 2.5.1-pre.1 hot fix on Async.js
### Issue
Accumulate is assuming first argument as a function and second argument as a context.
But, Luna send is calling response by sending first argument as an object and second argument as a string.

### Fix
Adding a check for second argument to cover the Luna send case.

DCO-1.1-Signed-Off-By: Kunmyon Choi kunmyon.choi@lge.com