BHV-12306: VideoPlayer does not unload video
### Issue
Can not unload video because it is set sourceComponents again on updateSource

### Fix
Not updateSource is checking existence of 'src' only. This makes one of src and sourceComponets set again.
In this case, when set src as "" updateSource is called but sourceComponents is not set null yet.

To make it safer, check the existence of source and existence of src/sourceComponents specifically.

DCO-1.1-Signed-Off-By: Kunmyon Choi kunmyon.choi@lge.com