BHV-12663: enyo.Video should not remove component block on create
### Issue
When developer adds child component with component block, it is removes on creation time.
Because, enyo.Video is updateSource on create and this makes remove all the components.

### Fix
Don't remove clientControls when updateSource is called from create.
The updateSource needs to be called to support sourceComponents in static.

DCO-1.1-Signed-Off-By: Kunmyon Choi kunmyon.choi@lge.com