CMB-1235: enyo.Video should also support components block to add "source"
### Issue
Latest enyo.Video is support sourceComponents to add source tags under the video tag.
Based on the general concept of enyo, it's better to support normal components block with defaultKind to be a video source.
In addition to this, we also should guarantee backward compatibility for sourceComponents.

### Fix
Look at the arguments on createComponent, and if the components block of enyo.Video has src or type then set these on attributes directly.
By doing this, we can have four ways to set source on enyo.Video.
- src:"http://vjs.zencdn.net/v/oceans.mp4"
- components: [{tag:'source', attributes: {src:"http://vjs.zencdn.net/v/oceans.mp4", type:"video/mp4"}}]
- components: [{src:"http://vjs.zencdn.net/v/oceans.mp4", type:"video/mp4"}]
- sourceComponents: [{src:"http://vjs.zencdn.net/v/oceans.mp4", type:"video/mp4"}]

DCO-1.1-Signed-Off-By: Kunmyon Choi kunmyon.choi@lge.com