ENYO-275: Properly initialize DataList.absoluteShowing
DataList maintains an internal absoluteShowing property and
observes changes to it, for the purpose of deferring certain
operations when the list is hidden.

Previously, DataList initialized absoluteShowing when the list was
created, setting it to value of the list's own "showing" property.
This meant it wasn't accurate if any of the list's ancestors
happened to be hidden at the time the list was created...and this,
in turn, meant that certain operations (e.g. initial generation of
pages and layout of child controls) that should have been deferred
happened immediately instead.

We now initialize on render instead, getting the value from a call
to getAbsoluteShowing().

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)