ENYO-307: DataList generation issue when list hidden...
...immediately after rendering.

By default, DataList doesn't render list contents immediately; it
renders them after a brief delay, to speed up the initial render
of the view containing the list.

It also waits to generate contents until the list is actually
showing, because it needs to measure the dimensions of the list
to generate the list pages correctly.

We had an issue in the edge case where a list was rendered and
then immediately hidden, because the initial check for
'absoluteShowing' would return true, but then the list would be
hidden by the time the async generation routine ran, and the bogus
measurements would cause the generation to go wrong.

To fix, we refactor things a bit to ensure that the check for
'absoluteShowing' always happens immediately before the generation
routine runs.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)