BHV-11799: Add new sample to test DataList / Scroller interaction
In addressing BHV-11799, I found it helpful to create a sample to
exhaustively test the interaction of DataList and DataGridList
with our various scroll strategies.

Submitting this as a separate pull request so that this sample can
more easily be used to test behavior both before and after the
changes made to fix BHV-11799.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)