ENYO-424: Prevent tap event when user presses to stop scrolling
When we stop scrolling in response to a 'down' event, we should
prevent a tap event from firing, on the assumption that the user's
intent was simply to stop the scroller.

To do so, I made the following changes:

* In gesture.js, extended the preventTap() API so that it can be
used on not only up events, but down events as well

* Changed our scroll strategies to stop scrolling on 'down'
(utilizing the above-mentioned preventTap() API), rather than on
'hold'. Implemented this functionality in ScrollStrategy,allowing
it to be inherited by subkinds.

* To support the above, also added support in ScrollStrategy for
onScrollStart, onScrollStop and isScrolling(). As a side benefit,
this makes for better API consistency across our scroll strategies.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)