BHV-19352: Fix controlsPerPage issue in vertical grid delegate
controlsPerPage in vertical delegate was recently updated to
invalidate cached page metrics. This was required to ensure that
pages were generated correctly after the list has been resized.

However, vertical grid delegate has been overriding controlsPerPage
in a way that is not compatible with the recent change, resulting
in calculation errors for DataGridList.

In this commit, we refactor the controls-per-page calculation so
that the vertical grid delegate can override just the portion it
needs to change, while preserving the rest of the logic.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)