ENYO-655: Removing modelsAdded optimization to fix pagination bugs
At some point we added an optimization in vertical delegate's
modelsAdded() method to avoid refreshing DataList when the models
being added all came later in the list than the currently rendered
pages.

This optimization is not currently working, causing issues like
ENYO-655 and ENYO-472.

I am still investigating a deeper fix for these problems that would
enable us to preserve the modelsAdded() optimization, but for now
it appears that removing the optimization is the simplest, lowest-
risk way of addressing the pagination issues we're seeing.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)