BHV-19877: Stabilize when remeasuring
We recently added a feature allowing users of a scroller to cue
a remeasurement of the scroller's bounds. This is useful when an
app or a component knows that the scroller's contents have changed,
since it allows the scroller to adjust things like thumb position
and size as needed in response to the content change.

Until now, however, remeasure() would not check to see if the
scroller's current scroll position was still valid. If the contents
have been reduced to the point where the current position is no
longer valid, the scroller should also stabilize itself to bring
its scroll position within bounds.

This issue was observed when working with Moonstone's data list
and data grid list; after removing models, the scroller's position
might be out of bounds. The list delegate was already calling
remeasure() in this case, but because remeasure() was not
stabilizing, the invalid scroll position was not corrected.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)