BHV-20938: Restore package filename when chunking
In dd96f39 (BHV-6022, "iLib divorce"), we changed the way that
mapped packages were represented as "chunks" in the walker.
Specifically, we were previously using the full path to the
package.js file, but started using just the path to the directory.

This change caused a problem for apps including packages with non-
standard package filenames (e.g. Moonstone's light-package.js);
the specific package filename would be stripped, and as a result
the package would be mapped to the default package.js file rather
than the specific variant the app was meant to include.

To fix, we simply revert to using the full path when adding mapped
packages as chunks. I have tested to ensure that this works with
both the ares-package CLI tool, and with the OE build process for
preparing on-device source.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)