precalculates scroll position offset on list render
Previous logic assumed that the offset would be calculated when the
first page was at the top. If the range isn't calculated until the
pages have been shifted for later indices, it calculated the offset
incorrectly.

This adds a little more burden to the list render phase which may be a
concern later.

Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)