Call adjust() on create, not rendered
The adjust() method sets a transform which controls the degree to
which the drawer is open. We were previously waiting until rendered
to initially set the transform, but by doing so we would sometimes
overwrite a value previously set.

Instead, we call adjust() on create, so that the initial (default)
transform will always be set first.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)