ENYO-1188: Accessibility: Apply "aria-label" to all enyo control which have a content
When content of control is set or chagned, aria-label should be also set
and changed. So I added the code of setting aria-label in contentChanged
function.
In addition, it needs the tabindex attribute to be focusable because
screen reader can read aria-label when dom is focused. And if dom has
tabindex attribute default outline is shown when it focused, so remove
the outline in case dom has tabindex attribute.

https://jira2.lgsvl.com/browse/ENYO-1188
Enyo-DCO-1.1-Signed-off-by: Jaewon Jang <jaewon98.jang@lgepartner.com>