ENYO-1125: Add "accessibilityHint" to enyo control
AccessibilityHint is used for additional information of control. If
accessibilityHint is set and content exists, screen reader reads
accessibilityHint with content when control is focused.
For example there is a button and if button clicked, image or video will
be played. In this case blind needs to know what will happen after
clicking button.
AccessibilityHint can explain the result of action and describe
something need to more information.

https://jira2.lgsvl.com/browse/ENYO-1125
Enyo-DCO-1.1-Signed-off-by: Jaewon Jang <jaewon98.jang@lgepartner.com>