ENYO-1124: Accessibility: Add "accessibilityAlert" to enyo control
The accessibilityAlert property is to cover alert role of aria
attribute.
Basically screen reader reads content or accessibilityLabel out when the
control is focused, but there are some cases that it should be read
something regardless focus. For example, popup, notifications, page
description and so on. If accessibilityAlert is set true, screen reader
will automatically reads content or accessibilityLabe when the control
is rendered or visible.
Refer

https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/ARIA_Techniques/Using_the_alert_role.

https://jira2.lgsvl.com/browse/ENYO-1124
Enyo-DCO-1.1-Signed-off-by: Jaewon Jang jaewon98.jang@lgepartner.com