Fix componentOverrides to work with modules
We filter out functions when mixing in componentOverride props, but
in the post-module world our convention is to pass constructors
(not strings) as values for 'kind'.

We therefore adjust our filter to allow functions only in the case
where they're passed as values for 'kind' (and add some comments to
explain the logic.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)