ENYO-1460: In NewDataList, constrain 'first'...
...such that the first rendered item is always the first item in
its own column or row. This ensures that we won't have partial rows
or columns unless / until we get to the final row or column in the
list and don't have enough items to fill it.

We do this to prevent odd behavior when 5-way navigating up and
down a list (using the moonstone/NewDataList subkind with Spotlight
support). In moonstone/NewDataList, we auto-scroll when navigating
by allowing focus to move naturally and scrolling the focused
element into view, which in turn causes previously virtualized
elements to be rendered into the DOM as needed. If the first off-
screen row (for example) is only partially rendered, focus can
"jump" laterally to the wrong item, because Spotlight can of course
only focus on already-rendered items.

We make the fix here, rather than in Moonstone, because it seems
like there is no case in which it's helpful to have partial rows
or columns, and there are likely other cases where it's harmful.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)