ENYO-1495: ScrollMath: Disregard boundary changes during overscroll
The boundary damping calculation in ScrollMath gives poor results
if the boundaries change while the scroller is in overscroll.

To fix, we take a snapshot of the boundaries when we enter
overscroll and use the snapshot for the damping calculation. We
discard the snapshot when we are no longer overscrolling.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)