ENYO-1461: In enyo/NewDataList, set first = 0 on `reset()`
Under certain circumstances after the underlying data had changed,
the first items in a list were not being rendered because we were
failing to reset the `first` property in `reset()`.

This issue was discovered and fixed when addressing the grab-bag
of other issues listed in ENYO-1461.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)